+++ docs/examples/opentracing/main.py
@@ -4,7 +4,7 @@ from rediscache import RedisCache
 
 from opentelemetry import trace
 from opentelemetry.exporter.jaeger import JaegerSpanExporter
-from opentelemetry.instrumentation import opentracing_shim
+from opentelemetry.shim import opentracing_shim
 from opentelemetry.sdk.trace import TracerProvider
 from opentelemetry.sdk.trace.export import SimpleExportSpanProcessor
