+++ shim/opentelemetry-opentracing-shim/MANIFEST.in
