+++ shim/opentelemetry-opentracing-shim/setup.cfg
