+++ shim/opentelemetry-opentracing-shim/src/opentelemetry/shim/opentracing_shim/util.py
